Move DataRequest to own file
